// web/app/components/workflow/types.ts
export const BlockEnum = {
  Start: 'start',
  StartPlaceholder: 'start-placeholder',
  End: 'end',
  Answer: 'answer',
  LLM: 'llm',
  KnowledgeRetrieval: 'knowledge-retrieval',
  QuestionClassifier: 'question-classifier',
  IfElse: 'if-else',
  Code: 'code',
  TemplateTransform: 'template-transform',
  HttpRequest: 'http-request',
  VariableAssigner: 'variable-assigner',
  VariableAggregator: 'variable-aggregator',
  Tool: 'tool',
  ParameterExtractor: 'parameter-extractor',
  Iteration: 'iteration',
  DocExtractor: 'document-extractor',
  ListFilter: 'list-operator',
  IterationStart: 'iteration-start',
  Assigner: 'assigner', // is now named as VariableAssigner
  Agent: 'agent',
  AgentV2: 'agent-v2',
  IntentExecutor: 'intent-executor',
  Loop: 'loop',
  LoopStart: 'loop-start',
  LoopEnd: 'loop-end',
  HumanInput: 'human-input',
  DataSource: 'datasource',
  DataSourceEmpty: 'datasource-empty',
  KnowledgeBase: 'knowledge-index',
  TriggerSchedule: 'trigger-schedule',
  TriggerWebhook: 'trigger-webhook',
  TriggerPlugin: 'trigger-plugin',
} as const

export type BlockEnum = (typeof BlockEnum)[keyof typeof BlockEnum]
// ...
export type ValueSelector = string[] // [nodeId, key | obj key path]

export type Variable = {
  variable: string
  label?:
    | string
    | {
        nodeType: BlockEnum
        nodeName: string
        variable: string
      }
  value_selector: ValueSelector
  value_type?: VarType
  variable_type?: VarKindType
  value?: string
  options?: string[]
  required?: boolean
  isParagraph?: boolean
}
// ...
export const InputVarType = {
  textInput: 'text-input',
  paragraph: 'paragraph',
  select: 'select',
  number: 'number',
  url: 'url',
  files: 'files',
  json: 'json', // obj, array
  jsonObject: 'json_object', // only object support define json schema
  contexts: 'contexts', // knowledge retrieval
  iterator: 'iterator', // iteration input
  singleFile: 'file',
  multiFiles: 'file-list',
  loop: 'loop', // loop input
  checkbox: 'checkbox',
} as const

export type InputVarType = (typeof InputVarType)[keyof typeof InputVarType]

export type InputVar = {
  type: InputVarType
  label:
    | string
    | {
        nodeType: BlockEnum
        nodeName: string
        variable: string
        isChatVar?: boolean
      }
  variable: string
  max_length?: number
  default?: string | number | boolean
  required: boolean
  hint?: string
  options?: string[]
  value_selector?: ValueSelector
  placeholder?: string
  unit?: string
  getVarValueFromDependent?: boolean
  hide?: boolean
  isFileItem?: boolean
  json_schema?: string | Record<string, any> // for jsonObject type
} & Partial<UploadFileSetting>
// ...
export const VarType = {
  string: 'string',
  number: 'number',
  integer: 'integer',
  secret: 'secret',
  boolean: 'boolean',
  object: 'object',
  file: 'file',
  array: 'array',
  arrayString: 'array[string]',
  arrayNumber: 'array[number]',
  arrayObject: 'array[object]',
  arrayBoolean: 'array[boolean]',
  arrayFile: 'array[file]',
  any: 'any',
  arrayAny: 'array[any]',
} as const

export type VarType = (typeof VarType)[keyof typeof VarType]
// ...
export const SupportUploadFileTypes = {
  image: 'image',
  document: 'document',
  audio: 'audio',
  video: 'video',
  custom: 'custom',
} as const

export type SupportUploadFileTypes =
  (typeof SupportUploadFileTypes)[keyof typeof SupportUploadFileTypes]

export type UploadFileSetting = {
  allowed_file_upload_methods: TransferMethod[]
  allowed_upload_methods?: TransferMethod[]
  allowed_file_types: SupportUploadFileTypes[]
  allowed_file_extensions?: string[]
  max_length: number
  number_limits?: number
}